# Optimize decode