--- a/src/sourcemap-codec.ts
+++ b/src/sourcemap-codec.ts
@@ -13,31 +13,30 @@ for (let i = 0; i < chars.length; i++) {
 }
 
 export function decode(mappings: string): SourceMapMappings {
-	let generatedCodeColumn = 0; // first field
-	let sourceFileIndex = 0;     // second field
-	let sourceCodeLine = 0;      // third field
-	let sourceCodeColumn = 0;    // fourth field
-	let nameIndex = 0;           // fifth field
-
 	const decoded: SourceMapMappings = [];
 	let line: SourceMapLine = [];
-	let segment: number[] = [];
+	const segment: SourceMapSegment = [
+		0, // generated code column
+		0, // source file index
+		0, // source code line
+		0, // source code column
+		0, // name index
+	];
 
-	for (let i = 0, j = 0, shift = 0, value = 0, len = mappings.length; i < len; i++) {
+	let j = 0;
+	for (let i = 0, shift = 0, value = 0; i < mappings.length; i++) {
 		const c = mappings.charCodeAt(i);
 
 		if (c === 44) { // ","
-			if (segment.length) line.push(segment as SourceMapSegment);
-			segment = [];
+			segmentify(line, segment, j);
 			j = 0;
 
 		} else if (c === 59) { // ";"
-			if (segment.length) line.push(segment as SourceMapSegment);
-			segment = [];
+			segmentify(line, segment, j);
 			j = 0;
 			decoded.push(line);
 			line = [];
-			generatedCodeColumn = 0;
+			segment[0] = 0;
 
 		} else {
 			let integer = charToInteger[c];
@@ -61,37 +60,27 @@ export function decode(mappings: string): SourceMapMappings {
 					if (value === 0) value = -0x80000000;
 				}
 
-				if (j == 0) {
-					generatedCodeColumn += value;
-					segment.push(generatedCodeColumn);
-
-				} else if (j === 1) {
-					sourceFileIndex += value;
-					segment.push(sourceFileIndex);
-
-				} else if (j === 2) {
-					sourceCodeLine += value;
-					segment.push(sourceCodeLine);
-
-				} else if (j === 3) {
-					sourceCodeColumn += value;
-					segment.push(sourceCodeColumn);
-
-				} else if (j === 4) {
-					nameIndex += value;
-					segment.push(nameIndex);
-				}
-
+				segment[j] += value;
 				j++;
 				value = shift = 0; // reset
 			}
 		}
 	}
 
-	if (segment.length) line.push(segment as SourceMapSegment);
+	segmentify(line, segment, j);
 	decoded.push(line);
 
 	return decoded;
+}
+
+function segmentify(line: SourceMapSegment[], segment: SourceMapSegment, j: number) {
+	// This looks ugly, but we're creating specialized arrays with a specific
+	// length. This is much faster than creating a new array (which v8 expands to a
+	// capacity of 17 after pushing the first item), or slicing out a subarray
+	// (which is slow).
+	if (j === 1) line.push([segment[0]]);
+	else if (j === 4) line.push([segment[0], segment[1], segment[2], segment[3]]);
+	else if (j === 5) line.push([segment[0], segment[1], segment[2], segment[3], segment[4]]);
 }
 
 export function encode(decoded: SourceMapMappings): string {