@jridgewell 2019-07-09
Combined with https://github.com/Rich-Harris/sourcemap-codec/pull/79, this nets us a 4-20x boost over v1.4.5:

```text
$ npx node@12 benchmark.js
v12.6.0
decode (v1.4.5) x 6,563 ops/sec ±1.30% (90 runs sampled)
decode (proposal) x 25,912 ops/sec ±2.06% (91 runs sampled)
Fastest is decode (proposal)

$ npx node@10 benchmark.js
v10.16.0
decode (v1.4.5) x 1,622 ops/sec ±2.49% (76 runs sampled)
decode (proposal) x 27,953 ops/sec ±4.11% (88 runs sampled)
Fastest is decode (proposal)

$ npx node@8 benchmark.js
v8.16.0
decode (v1.4.5) x 6,029 ops/sec ±1.85% (92 runs sampled)
decode (proposal) x 17,079 ops/sec ±1.12% (92 runs sampled)
Fastest is decode (proposal)

$ npx node@6 benchmark.js
v6.17.1
decode (v1.4.5) x 1,313 ops/sec ±3.83% (78 runs sampled)
decode (proposal) x 21,032 ops/sec ±0.75% (93 runs sampled)
Fastest is decode (proposal)
```