@jridgewell 2019-07-12
Ping @Rich-Harris, anything I can do to get this in?