@Rich-Harris 2020-01-16
Thank you! This is great — released 1.4.8. Apologies for the delayed response, my inbox was a disaster zone until recently, so I missed the pings.